Removed plugin contributions
